Re-organize the tensorize test cases, and add a new case for none-reset mode.
